# backend/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /api/v1/novel/projects/{projectId}/video-script:
        post:
            tags:
                - NovelService
            description: 生成短视频分镜脚本
            operationId: NovelService_GenerateVideoScript
            parameters:
                - name: projectId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/novel.v1.GenerateVideoScriptRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/novel.v1.GenerateVideoScriptResponse'
...
    /api/v1/projects/{projectId}/video-scripts:
        get:
            tags:
                - VideoScriptService
            description: 获取视频脚本列表
            operationId: VideoScriptService_ListVideoScripts
            parameters:
                - name: projectId
                  in: path
                  required: true
                  schema:
                    type: string
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.video_script.v1.ListVideoScriptsResponse'
    /api/v1/video-scripts/generate:
        post:
            tags:
                - VideoScriptService
            description: 生成短视频脚本
            operationId: VideoScriptService_GenerateVideoScript
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.video_script.v1.GenerateVideoScriptRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.video_script.v1.GenerateVideoScriptResponse'
    /api/v1/video-scripts/{baseScriptId}/variants:
        post:
            tags:
                - VideoScriptService
            description: 生成平台变体
            operationId: VideoScriptService_GeneratePlatformVariants
            parameters:
                - name: baseScriptId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.video_script.v1.GeneratePlatformVariantsRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.video_script.v1.GeneratePlatformVariantsResponse'
    /api/v1/video-scripts/{scriptId}:
        get:
            tags:
                - VideoScriptService
            description: 获取视频脚本详情
            operationId: VideoScriptService_GetVideoScript
            parameters:
                - name: scriptId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.video_script.v1.GetVideoScriptResponse'
        delete:
            tags:
                - VideoScriptService
            description: 删除视频脚本
            operationId: VideoScriptService_DeleteVideoScript
            parameters:
                - name: scriptId
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.video_script.v1.DeleteVideoScriptResponse'
    /api/v1/video-scripts/{scriptId}/optimize:
        put:
            tags:
                - VideoScriptService
            description: 优化短视频脚本
            operationId: VideoScriptService_OptimizeVideoScript
            parameters:
                - name: scriptId
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.video_script.v1.OptimizeVideoScriptRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.video_script.v1.OptimizeVideoScriptResponse'
...
components:
    schemas:
        api.video_script.v1.DeleteVideoScriptResponse:
            type: object
            properties:
                success:
                    type: boolean
            description: 删除视频脚本响应
        api.video_script.v1.GeneratePlatformVariantsRequest:
            type: object
            properties:
                baseScriptId:
                    type: string
                targetPlatforms:
                    type: array
                    items:
                        type: string
            description: 生成平台变体请求
        api.video_script.v1.GeneratePlatformVariantsResponse:
            type: object
            properties:
                variants:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.video_script.v1.VideoScript'
            description: 生成平台变体响应
        api.video_script.v1.GenerateVideoScriptRequest:
            type: object
            properties:
                projectId:
                    type: string
                chapterId:
                    type: string
                chapterTitle:
                    type: string
                chapterContent:
                    type: string
                platform:
                    type: string
                duration:
                    type: integer
                    format: int32
                style:
                    type: string
                requirements:
                    type: string
            description: 生成短视频脚本请求
        api.video_script.v1.GenerateVideoScriptResponse:
            type: object
            properties:
                videoScript:
                    $ref: '#/components/schemas/api.video_script.v1.VideoScript'
            description: 生成短视频脚本响应
        api.video_script.v1.GetVideoScriptResponse:
            type: object
            properties:
                videoScript:
                    $ref: '#/components/schemas/api.video_script.v1.VideoScript'
            description: 获取视频脚本详情响应
        api.video_script.v1.ListVideoScriptsResponse:
            type: object
            properties:
                scripts:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.video_script.v1.VideoScript'
                total:
                    type: integer
                    format: int32
            description: 获取视频脚本列表响应
        api.video_script.v1.OptimizeVideoScriptRequest:
            type: object
            properties:
                scriptId:
                    type: string
                requirements:
                    type: string
            description: 优化短视频脚本请求
        api.video_script.v1.OptimizeVideoScriptResponse:
            type: object
            properties:
                videoScript:
                    $ref: '#/components/schemas/api.video_script.v1.VideoScript'
            description: 优化短视频脚本响应
        api.video_script.v1.VideoHooks:
            type: object
            properties:
                opening:
                    type: string
                climax:
                    type: string
                ending:
                    type: string
            description: 视频钩子点
        api.video_script.v1.VideoScene:
            type: object
            properties:
                index:
                    type: integer
                    format: int32
                duration:
                    type: integer
                    format: int32
                shotType:
                    type: string
                visualDescription:
                    type: string
                narration:
                    type: string
                subtitle:
                    type: string
                soundEffects:
                    type: array
                    items:
                        type: string
                transition:
                    type: string
                keyElements:
                    type: array
                    items:
                        type: string
            description: 视频分镜场景
        api.video_script.v1.VideoScript:
            type: object
            properties:
                id:
                    type: string
                projectId:
                    type: string
                chapterId:
                    type: string
                title:
                    type: string
                duration:
                    type: integer
                    format: int32
                platform:
                    type: string
                style:
                    type: string
                scenes:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.video_script.v1.VideoScene'
                hooks:
                    $ref: '#/components/schemas/api.video_script.v1.VideoHooks'
                hashtags:
                    type: array
                    items:
                        type: string
                description:
                    type: string
                status:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
            description: 短视频脚本
...
        novel.v1.GenerateVideoScriptRequest:
            type: object
            properties:
                projectId:
                    type: string
                chapterIds:
                    type: array
                    items:
                        type: string
                options:
                    $ref: '#/components/schemas/novel.v1.VideoScriptOptions'
            description: 短视频分镜脚本相关消息
        novel.v1.GenerateVideoScriptResponse:
            type: object
            properties:
                scenes:
                    type: array
                    items:
                        $ref: '#/components/schemas/novel.v1.VideoScene'
...
        novel.v1.VideoScene:
            type: object
            properties:
                screenIndex:
                    type: integer
                    format: int32
                text:
                    type: string
                suggestedBgmTag:
                    type: string
                suggestedImageTag:
                    type: string
                ttsVoice:
                    type: string
                notes:
                    type: string
        novel.v1.VideoScriptOptions:
            type: object
            properties:
                scenesPerChapter:
                    type: integer
                    format: int32
                platform:
                    type: string
                voiceType:
                    type: string
...